Fix warnings (#145)
* Remove unused 'hardlimit' function parameter

* Fix conversion of string to char*

* Apply clang-format

* fix warning : removing unused local variable

* fix warning : no need for a getstarttime() function

* Fix warnings : matrix tests and predicates

A TSNumbersPredicate should not be used to save a Matrix<double, T>. It is not meant for this.

* fix warning : runtime-info --> type mismatch fixing

* fix warnings : on loading parameters from disk

* fix warning : hydro/series.cpp - conversion and mismatch

* fix warning : cbuilder.cpp - conversions

* fix warning : constraint builder - conversion to uint

* fix warning : notebook - upgrading outline and inside color style constant for a shape

* fix warning : notebook - upgrade wxWidgets constants for fonts, filling and outlining areas.

* fix warning : unfeasible problems - removing unused exception variables

* fix warning : removing mismatches between signed/unsigned due to a '<'

* fix warning : pop messages - removing a useless reinterpret_cast(...)

* fix warning : notebook - upgrade wxWidgets constants for outlining areas.

* fix warning : removing unused variables + adding a 'break' in a switch

* fix warning : removing unused argument

* fix warning : enum as boolean

* fix warning : survey results - unused vars + switch uncompletee

* fix warning : reordering member initialization

* fix warning : reordering member initialization

* fix warning : switch uncomplete

* fix warnings : remove of unused var

* fix warning : reordering member initialization

* fix warning : fixing comparison of integer of different signedness

* fix warning : switch uncomplete

* fix warning : making Matrix destructor virtual (Matrix used as polymorphic object)

* fix warning : enum as boolean

* fix warning : enum as boolean

* fix warning : converting uint into int

* fix warning : reordering member initialization

* fix warning : reordering member initialization

* fix warning : moving a 'break' in a switch to its right place

* fix warning : parameters.cpp - several things corrected

* fix warning : file constraint.cpp

* fix warning : file container.cpp

* fix warning : removing an unused argument

* fix warning : reordring members in constuctor + removing unused var

* fix warning : removing unused var

* fix warning : removing unused argument

* fix warning : reordring members in constuctor + removing unused arg

* [skp ci] fix warning : removing encoding problems

* [skip ci] fix warning : removing unnecessary comment

* [skip ci] fix warning : removing encoding problems

* [skip ci] fix warning : correcting wrong integer comparison

* [skip ci] fix warning : restoring a useful warning

* [skip ci] fix warning : resolving encoding problem

* [skip ci] fix warning : parameters - fixing enum value comparison

* fix warning : restoring a useful warning

+++ src/solver/constraints-builder/grid.hxx
@@ -137,7 +137,7 @@ uint Grid<NodeT>::getNumberOfConnectedComponents()
         cVec[cIT->second] = true;
     }
 
-    return cVec.size();
+    return (uint)(cVec.size());
 }
 
 template<class NodeT>
